+++ controllers/addController.js
@@ -1,6 +1,14 @@
 const asyncHandler = require("express-async-handler");
 const { body, validationResult } = require("express-validator");
-const { getCategories, insertCategory } = require("../db/queries");
+const {
+  getCategories,
+  getCategory,
+  getItem,
+  insertCategory,
+  insertItem,
+  updateItem,
+} = require("../db/queries");
+const foo = require("../db/queries");
 
 const validateCategory = [
   body("category")
@@ -13,12 +21,12 @@ const validateCategory = [
 const validateItem = [
   body("category")
     .trim()
-    .isLength({ min: 3, max: 20 })
     .isAlpha()
-    .optional({ values: "falsy" })
+    .isLength({ min: 3, max: 20 })
     .withMessage(
       "Category input is optional. However, the value must be between 3 and 20 letters long."
-    ),
+    )
+    .optional({ values: "falsy" }),
   body("item")
     .trim()
     .isLength({ min: 3, max: 30 })
@@ -27,14 +35,17 @@ const validateItem = [
     .trim()
     .isNumeric()
     .isLength({ min: 10, max: 12 })
+    .escape()
     .withMessage("UPC must be between 10 and 12 digits long."),
   body("quantity")
     .trim()
     .isInt({ min: 0, max: 999999 })
+    .escape()
     .withMessage("Quantity must be between 0 and 999,999."),
   body("price")
     .trim()
     .isFloat({ min: 0.01, max: 999999 })
+    .escape()
     .withMessage("Price must be between 0.01 and 999,999."),
 ];
 
@@ -77,7 +88,7 @@ const addController = {
       const { category } = req.body;
 
       try {
-        await insertCategory(category);
+        await insertCategory(req.body);
         const categories = await getCategories();
         res.render("addCategory", {
           title: "Add Category",
@@ -103,9 +114,8 @@ const addController = {
     validateItem,
     asyncHandler(async (req, res) => {
       const errors = validationResult(req);
-
+      const categories = await getCategories();
       if (!errors.isEmpty()) {
-        const categories = await getCategories();
         const localErrors = errors
           .array()
           .reduce((accumulator, currentError) => {
@@ -113,11 +123,64 @@ const addController = {
             return { ...accumulator, [path]: { value, msg } };
           }, {});
 
+        // console.log(errors);
+        console.log("localErrors");
+        console.log(localErrors);
         return res.status(400).render("addItem", {
           title: "Add Item",
           errors: { ...localErrors },
           inputs: { ...req.body },
           categories,
+        });
+      }
+
+      console.log("postAddItem running...");
+      console.log(req.body);
+      try {
+        // categoryExists and upcExists returns 0 or 1
+        // 0 Indicates no result and 1 indicates existing result(s)
+        const { rowCount: categoryExists } = await getCategory(req.body);
+        const { rowCount: upcExists } = await getItem(req.body);
+        console.log(categoryExists);
+        console.log(upcExists);
+
+        if (categoryExists && !upcExists) {
+          console.log("category exists and upc does NOT exist");
+          // Insert item
+          await insertItem(req.body);
+        } else if (!categoryExists && upcExists) {
+          console.log("category does NOT exist and upc exists");
+          // Insert category
+          // Update item
+          // await insertCategory(req.body);
+          // await foo.updateItem(req.body);
+          // Redirect user to an update page?
+        } else if (categoryExists && upcExists) {
+          console.log("category and upc both exist");
+          // Update item
+          // await foo.updateItem(req.body);
+          // Redirect user to an update page?
+        } else {
+          console.log("category and upc both do NOT exist");
+          // Insert category
+          // Insert item
+          await insertCategory(req.body);
+          await insertItem(req.body);
+        }
+
+        // How to let the user know the item was successfully added?
+        // res.redirect("/add/item");
+        res.status(200).render("addItem", {
+          title: "Add Item",
+          inputs: { ...req.body },
+          categories,
+        });
+      } catch (err) {
+        console.log("err:");
+        console.log(err);
+        return res.status(400).render("addItem", {
+          title: "Add Item",
+          inputs: { ...req.body },
         });
       }
     }),