+++ controllers/itemsController.js
@@ -1,0 +1,8 @@
+const asyncHandler = require("express-async-handler");
+const { getItems } = require("../db/queries");
+
+const itemsController = {
+  getItems: async () => {},
+};
+
+module.exports = itemsController;